gitconfig: Remove argument "-a" from alias "b"
The alias "b" is an abbreviation of the command "branch". Before this
change the argument "-a" was always passed into the command. This
severely restricted the use of the alias.

In order to use the alias freely, no arguments should be passed into it.
That way any argument of the "branch" command can also be used on the
alias.