gitconfig: Add alias "ignore-all"
Sometimes it happens that it's not possible to switch branches because
an ignored file has changes which would be overwritten by checkout. In
these cases it's necessary to stash the ignored files. I usually do it
like this:

$ git unignore-all
$ git stash
$ git switch other-branch
$ git stash pop

Thereafter I want to ignore all the files again. To make this part
easier the new alias "ignore-all" can be used which ignores all modified
files in the working tree.