emacs: Set value of variable `fill-column` to 80
It specifies when automatic line wrapping should happen.

The default value of `fill-column` is 70 but it is a bit short in my
opinion. The de facto standard is a text width of 80 characters since
longer lines cannot be displayed properly on most terminals.

For this reason use a value of 80.