git-check-repos: do not escape paths
Prior to this change, paths containing backslashes could have been
interpreted incorrectly. For example if a path contained the sequence
"\n" it would have been interpreted as a newline character.

To ensure `find` doesn't interpret any backslashes, the option "-print0"
is used in combination with `xargs` option "-0". In this case `find`
prints the path to stdout as is followed by a NUL character. To ensure
`read` does not treat backslashes in any special way, the option "-r" is
applied.