git-check-repos: speed up searching of repos
The old solution was pretty naive, after it found a ".git/" directory it
didn't stop but happily searched the whole repository if there could be
another one. This made the lookup process pretty slow.

Do not traverse a directory any further if it is known to be a Git
repository (contains ".git/") to improve the performance.