bashrc: try to source Git's bash-completion file
Since commit 0f9812e (bash_aliases: Add alias to abbreviate "git",
2020-07-07) the function `__git_complete` is used. This function is
defined in "/usr/share/bash-completion/completions/git", thus, this file
has to be sourced.

It may happen that this file is not present on the machine, e.g. Git is
not installed. To avoid causing a "command not found" error, check the
existence of these functions before using them.