emacs: allow using tab in terminal
A terminal cannot distinguish between CTRL-i and the tab key which means
if CTRL-i is bound then tab is doing the same. I just want to insert a
simple tab character, thus, it must not be bound.

Also I'm using CTRL-i very rarely so it doesn't hurt disabling it on the
terminal.