sway: change keybinding to open terminal
In Org-mode "M-return" can be used to insert new headings. That's a very
helpful command.

Prior to this change it couldn't be used in Emacs since it was used by
sway. Now, to open a terminal Ctrl+Alt+t is used instead which is also a
common keybinding in other desktop environments.