emacs: add package "Magit"
I use Git on the CLI for a long time and I didn't feel I was missing
out, but on the internet everybody is praising this package. So I gave
it a shot for a few weeks. My first impression is that some things can
be done really fast.

There is so much more to learn and maybe I'll switch my complete Git
workflow over to Magit in the future. We will see.

Furthermore the variable `fill-column` has been set to 72 when writing
commit messages to comply with the 50/72 rule.