emacs: hide some visual elements to have more space
The laptop which I'm using at home has a relative low resolution so
screen space is precious.

Furthermore I'm not using either the menu bar nor the toolbar. I try to
rely only on keybindings. Get rid of them.

The scrollbars I do not need either but I like that they give a visual
indication at which location in the document I'm in. But having them
enabled causes that many lines are wrapped when two windows are
opened. This bugs me even more, so hide them as well.