sway: use Super key as modifier key
Prior to this change Sway used the Alt key as its modifier key.

When using Emacs many keybindings require the usage of the Meta key,
a.k.a. the Alt key. As a result many keybindings of Sway conflicted with
those of Emacs. To avoid those conflicts use the Super key instead.

I liked the idea of using Alt as my modifier key because I was familiar
using it on macOS at work but in combination with Emacs it doesn't make
much sense to keep it.