Revert "sway: change keybinding to open terminal"
This keybinding was changed since it conflicted with "M-RET" in Org-Mode
to insert new headings. But since the modifier key of Sway has been
changed from Alt to Super in 67ff4b9 (sway: use Super key as modifier
key, 2022-04-09), we can use the old keybinding again. And to be honest,
the old one is just easier to press.

This reverts commit b0b8d63dcc19783bf2f94bfbcff6c00364736904.