emacs: optimize installation of packages
Prior to this change the installation code of each package was kinda
similar. First check if the package is already installed, if not refresh
the package archive and then install the package.

With this approach it could happen that the package archive would be
updated multiple times in a row if multiple packages were missing. This
isn't necessary however it increases the startup time of Emacs.

This logic is now moved into its own function so that it can easily be
reused. Also the package archive will only be updated once even if
multiple packages are missing.