emacs: show tab bar
I made some adjustments to the tab bar so it matches my taste visually.
For one, I removed the new tab button as well as the close button since
I'm only using the keybindings anyway and the buttons don't look good at
all.

Furthermore, I adjusted the coloring of the tab bar and the tabs
themselves. Now the tab bar and the inactive tabs are in the same color
but a lighter one then the background color of the buffer. The active
tab has the same as the background color of the buffer.