emacs: add keybinding to instant commit
In some projects (e.g. my notes) I don't care about the commit message.
In such a case I just want to commit the changes, that's when the
instant commit comes into play. It just commits the staged changes and
includes the filenames of the changed files in the commit message.