emacs: highlight uncommitted changes
This functionality is provided by the package "diff-hl". I configured it
so that it only highlights unstaged changes. I like this behavior
because that way I only see changes which I haven't approved for commit
yet. Also it was necessary to add two hooks so that operations in Magit
are populated to diff-hl.

By default changes are shown in the fringe. However the fringe is only
shown on graphical displays, as a fallback the margin should be used
instead when Emacs is run in the terminal. Unfortunately I haven't found
a solution yet which I indicate by the TODO comment.