emacs: remove Evil package
I was a long-time Vim user before using Emacs. Because of this, at the
beginning of my Emacs career, I was using evil-mode. To familiarize
myself more with Emacs I've switched to the default keybindings.

At the moment I'm not sure what keybindings I should be using in the
long run. The problem I see with Evil is that many modes are just
designed for the default keybindings, using them with Evil requires
additional work.

Maybe I'll come back some day.