emacs: do not show images in EWW by default
Most of the time I'm browsing in EWW with a really slow internet
connection (64 kbits/s). To save bandwidth only the most important data
should be loaded.

Furthermore, most of the time without images the buffer is much more
readable. The images are often just to large or they don't provide any
relevant information to the text because they are only part of the
layout or the like.