bash/aliases: remove alias "rm"
With the option "-I" the command "rm" prompts before removing more than
three files or when removing a directory recursively.

In a sense this is a good thing because that way some mistakes can be
prevented. But I find myself more often using the "-f" option (force) to
avoid getting such prompts because I find them a bit annoying.

Furthermore, I probably shouldn't take it for granted that I'm prompted
before removal. There are systems where my dotfiles are not installed
and thus I cannot rely on this behavior.