bashrc: colorize output of "man"
The pager "less" can be used to set distinct colors for certain types of
texts. That way I can color the bold and underlined text in the output
of "man".

I chose the color green for bold text since it matches the footer of
tmux even though I'm not a fan of this color. Well, I should change the
color of the tmux footer some day.