install.sh: don't try to create already existing symlinks
Prior to this change, the first time the script was run, it created all
symlinks. The second time though, the user was asked if he/she wants to
overwrite them. This was rather annoying. Most of the time I piped the
script through `yes` to skip these prompts.

Now the script checks if a symlink already exists and points to the
intended file, if so nothing is done which means no prompt is shown.