install.sh: use blacklist instead of whitelist
The new logic simplifies the maintenance of this file. That's because
it's much more likely that I add a new file to this repository which
should be symlinked instead of being ignored.

The determination of the base directory has been moved to a variable so
that it can be reused. Furthermore, it was necessary that `realpath` is
called with "$0", otherwise a incorrect directory would be returned if
install.sh were a symlink in a different directory.