emacs: don't show relevant diff when committing
The setting commit.verbose is enabled in my Git configuration. As a
result the diff of the commit is already shown in the buffer where I'm
editing the commit message. Thus, I don't have a need for this feature.

Also it sometimes annoys me when the current non-Magit buffer is
replaced by the diff of the commit. Maybe I still wanted to have that
buffer open!?