emacs: add command to prettify JSON from clipboard
Sometimes I have to examine a JSON response of a server in the
development tools of a web browser. Usually the JSON response is just a
single line and therefore very hard to read.

In such case the new command comes in handy: Just copy the JSON response
to your clipboard. After invoking the new command, Emacs automatically
prettifies the JSON data and displays it in a new buffer.