emacs: only highlight whitespace before tab
Files which use spaces to indent lines are not nice to look at if all
spaces are visualized. This is especially nasty in Emacs Lisp files.
For this reason spaces are not visualized anymore.

In my opinion spaces are never needed before a tab. Therefore, depending
on the `indent-tabs-mode`, the spaces or tabs are highlighted in such a
case.