emacs: use Modus themes
Look-wise I find the Solarized Selenize Black theme not bad but the
Modus themes are just much more customizable. Also they allow toggling
between light and dark themes very easily. Furthermore version 4, which
I'm using now, will also be included in Emacs 30 by default.

I had to rewrite the coloring of the box used by the "tab-bar-tab" and
"tab-bar-tab-inactive" faces. For some reason, the color was not set
correctly when Emacs was started as a daemon. In that case the function
`face-attribute` returned just "unspecified" instead of the correct
color.