emacs: configure conf-, prog- and text-mode separately
Prior to this change, the same hook was called for all three kinds of
major modes (and their descendants). While testing different stuff, I
sometimes encountered new settings which are only relevant for some of
these modes.

To make it easier to use different settings for each kind in the future,
distinct functions are used now.