bash/prompt: indicate if HTTP proxy is set
At work we have a HTTP proxy in use. By default I don't want to send my
traffic to it. But when I do, an indicator should be shown.

The function get_proxy() has to return the exit status of the previous
command. If we didn't do that the exit status would be overriden by
get_proxy(). We don't want this because get_exit_status_color() wants to
know the exit status of the actual command.