emacs: use own function for `git-commit-mode-hook`
A dedicated function can easier be tested as a lambda. That's because
one only has to re-evaluate the function and then can immediately test
it which cannot be done with a lambda.

Also, as more functionality is added to the hook, the lambda expression
will get more and more unreadable.