git-check-repos: rename func. to find_git_repo_paths()
In the future it should be possible to provide a file which contains a
list of Git repositories. This should be an alternative to the current
behavior where the file system is search for Git repositories.

Rename the function to make it clearer that the Git repositories are
looked up. This is a preparation step for the following commit where we
want to use the old name for another function.