bash: append commands immediately to history file
By default commands are written to the history file after the bash
session ended but this comes with two disadvantages:

1. When multiple bash sessions are opened simultaneously, only the
commands of last session are stored in the history file. That's because
the last session overwrites the stored commands of the other sessions.

2. If an additional session is opened, the commands entered in sessions
which are not closed yet, are not available.

This commit ensures that commands are stored in the history file
immediately after execution. This solves the issues mentioned above.