bash: enable fzf-powered key bindings in the terminal
The default key binding (CTRL-R) to search the history is overwritten,
instead it allows to fuzzy search the history. This is the most helpful
key binding in my opinion.

Furthermore, two new key bindings are added: CTRL-T which inserts the
selected file path and ALT-C which changes to the selected directory.