bash: only source fzf key bindings if available
Since commit 4e2f303 (bash: only source fzf key bindings if terminal is
not dumb, 2023-05-21) this file is only sourced if the terminal is not
dumb. In that change it was forgotten to check if this file exists on
the system though. If this file were missing, it would fail. This commit
fixes this issues.