bash: remove unwanted fzf key bindings
The fzf script binds several keys. One of them is "M-c" which allows to
cd to a directory which one can select via fzf.

The problem is that the terminal cannot distinguish between "M-c" and
"C-[ c". Since readline is configured to use Vi-like key bindings, I use
the second key binding quite regularly or to be precise "C-[ c c". It
allows me to clear the current line and switch from normal mode to
insert mode. To be able to use "C-[ c" as wanted, I inhibit the script
to bind anything by itself.

Actually, from this script the only key binding I care about is "C-r". I
copied the definition of it over from the original script.

I think there should be an official way to avoid binding these keys. One
day I should open an issue in GitHub to address this issue.