bash: only bind __fzf_history__ if available
Since commit abfa3f8 (bash: remove unwanted fzf key bindings,
2023-05-24) we bind the keys of fzf ourselves. But the file is only
sourced if it exists which means it's possible that __fzf_history__ does
not exist. When then one tries to use C-r to search the history, he/she
gets following error:

  bash: __fzf_history__: command not found

To fix this issue, only bind the fzf command if it is available.