emacs: use own command to un-/comment lines
I just want to be able to un-/comment the current line or if a region is
active, to un-/comment the whole region. The commands `comment-dwim` and
`comment-line` which come with Emacs do not meet my requirements.

The command `comment-line` goes in the right direction but has two
problems:

1. It moves point one line further down after un-/commenting.
2. When using Evil mode and a region is active, it also acts upon the
   next line which isn't part of the region anymore.

Well, to do it properly I have to write my own command.