emacs: hide visual elements in early-init.el
The early init file is read before the GUI is initialized, the normal
init file afterwards. The visual elements should be disabled in the
early init file because that way it can be ensured that they will never
be initialized.

Prior to this change, the visual elements were initialized before the
normal init file was read. But when the normal init file was read, it
disabled them. One could observe this behavior: When Emacs was started
the menu bar was shown for a short amount of time and then disappeared.

This change ensures that the menu bar isn't shown at all at startup.