emacs: simplify navigation in Man mode
By default the keys "n" and "p" are used to move sections, but I don't
use them as frequently as next-line and previous-line. Therefore the
default key bindings are moved to "N" and "P" respectively and next-line
and previous-line are used instead.

This change is similar to ce70d96 (emacs: simplify navigation in eww
mode and Info mode, 2022-07-12).