bash: don't save "exit" in the history file
The following situation happens to me quite frequently:

I open ansi-term in Emacs, enter a command and examine its output and
close it again. Then I continue doing my stuff in Emacs, but at some
point I want to re-run the last command from the terminal. To do so, I
spawn a new ansi-term and type "C-] k" (I use the Vi key bindings) to
get the last command. It is "exit", though. Do I remember this fact, no.
My fingers are doing this so fast that I realize this only after I
pressed enter. And the terminal is closed.

Fortunately this can be fixed quite easily. Simply do not store the
"exit" command in the history file.