git: fix colors used by diff-so-fancy
If a terminal only supported 8 colors, then the configuration wasn't
working. That's because colors out of this range were used.

This commit moves the definition of these options to the .bashrc file.
That way it's possible to set the colors dynamically, depending on the
amount of colors supported by the current terminal.