emacs: change highlighting of symbol at point in eglot
Among others this face highlights all occurrences of a variable at
point. By default these occurrences are displayed in a bold face, but in
my opinion this doesn't stand out enough. For this reason I changed it
so that the background is highlighted.

This is done similarly in VSCode where I was inspired from.