emacs: add comments to a key binding
In this file we have another key binding which removes the local key
binding so that the global one is used. There it's already explained by
a comment. Use the same comment here as well.

Also a comment was added to describe which command is by default bound
in elisp-mode.