emacs: don't display output of async commands
It's possible to run commands asynchronously using `shell-command`. By
default the output of these commands is displayed in a separate buffer.

I find this distracting. If I run a command in the background then it
should do its thing *in the background*.

Now the buffer isn't displayed anymore but it's still created in the
background and can be accessed if needed.