emacs: allow using `xref-find-definitions` in Normal state
The key sequence "M-." (xref-find-definitions) is globally bound by
Xref. It is a pretty handy key binding to find the definition of an
identifier at point. Unfortunately Evil mode overwrites this key binding
in Normal state with its own command.

Since I have no use for the command provided by Evil, it is removed from
the keymap so that `xref-find-definitions` can be used instead.