emacs: ensure ElDoc uses a single line in minibuffer
If Eglot is enabled in a buffer, most identifiers provide its own doc
string and ElDoc shows them in the minibuffer.

By default ElDoc resizes the minibuffer to show doc strings which are
larger then one line. I find this pretty distracting because when I'm
navigating through a file many different doc strings are shown and the
minibuffer resizes itself many times.

If I want to see the whole doc string, I can simply use the command
`eldoc-doc-buffer` which is bound to "C-h ." in Eglot.