emacs: specify `magit-pop-revision-stack-format`
This variable controls how the command `magit-pop-revision-stack`
inserts a revision.

The format of this variable is (POINT-FORMAT EOB-FORMAT INDEX-REGEXP).
EOB-FORMAT allows to specify which text should be inserted at the end of
the buffer and INDEX-REGEXP is used to number the revisions. I don't
care about both of them, that's why they aren't provided.

The value of POINT-FORMAT is determined by using following command:

  git show --format=POINT-FORMAT

According to the man page git-show(1) the built-in format "reference"
expands to:

  <abbrev-hash> (<title-line>, <short-author-date>)

It also states that this format is used to refer to another commit in a
commit message.

This is exactly my use case. While I'm writing a commit message I often
encounter myself wanting to refer to another commit. In such case the
command `magit-pop-revision-stack` comes handy, especially now using the
proper format.

As a side note, this format is also used in other popular Open-source
projects such as Git.