emacs: allow pixelwise resize of frames
At work I have to use Windows 11 where I make use of WSLg to run Emacs.

Windows has this nice feature which allows to snap windows to the left
or right of the screen. However, the problem is when I try to snap a
Emacs frame, the frame is not using all available pixels on the screen.
The reason for this is the variable `frame-resize-pixelwise` which has
to be set to t.