emacs: use RET to jump to function in profiler
In the profiler's report buffer by default RET toggles the visibility of
the call tree. In my opinion this is not very intuitive, it should just
jump to the functions definition. This commit adds such key binding. To
toggle the visibility of the call tree, TAB can still be used.