bash/prompt: improve bash prompt in rare cases
A space is used to separate the branch name from the proxy indicator.
This space though was always part of the bash prompt even if the proxy
indicator wasn't shown.

In the rare case when no proxy indicator is shown and the first line of
the bash prompt has exactly the same number of characters as the width
of the terminal, an additional blank line was printed. That's because
the terminal needed one extra character for that space.

This change ensures that the space is only added when needed to avoid
the problem described above.