sway: use i3blocks as status_command
I like the simplicity of swaybar and don't need all the bells and
whistles which for example Waybar [0] provides. However the problem with
swaybar is, it's very hard to customize on its own.

i3blocks describes itself as "the hacker-friendly status_command for
Sway and i3". [1] With it it's possible to customize swaybar by writing
own scripts.

I like this flexible approach: It gives me the freedom to do whatever I
want with my status bar. That's the reason why I decided against using
i3status. [2]

For now the status bar shows the same information as before: the current
date and time. My current configuration of i3blocks is a trimmed-down
version of the default i3blocks configuration file. [3]

[0]: https://github.com/Alexays/Waybar
[1]: https://github.com/vivien/i3blocks
[2]: https://github.com/i3/i3status
[3]: https://github.com/vivien/i3blocks/blob/master/i3blocks.conf