emacs: use own keymap for leader key
Using your own keymap for the leader key is more flexible. It allows
changing the key binding of the leader key in a single place.
Furthermore the keymap can be reused for multiple key bindings.

I plan to use SPC as my leader key in Normal state of Evil mode. This is
the first step to do so.