emacs: truncate lines in Grep and Occur mode
Both Grep and Occur can be used to search for specific lines in a buffer
or file. If a matching line does not fit in the current window then it
should be truncated. That's because it makes analyzing the search
results for me easier. If there are wrapped lines in between, the whole
buffer looks just cluttered in my opinion.