emacs: allow opening Org agenda using a single key press
Most of the time I want to see the agenda for the current week but
pressing "C-c a a" feels just to cumbersome to do so. That's why it is
now possible to open it using F12.

I keep the key binding "C-c a" since it allows me to choose a different
agenda view if I want to.