i3blocks: consider all batteries of device
There are devices such as my Thinkpad T470s which have multiple
batteries. To know the proper remaining capacity, all these batteries
have to be checked.

The property `command` has been set in a way that it is possible to just
add a script with the name of the particular blocklet to the directory
.config/i3blocks/scripts/. The property `command` doesn't have to be
configured per blocklet anymore but if needed it is still possible like
it is done with the blocklet "time".

This works because i3blocks exports any property of a blocklet as an
environment variable. In this case `$name` is the name of the particular
blocklet.