Set some colors depending on env. variable "BACKGROUND"
Some colors used by Git and Vim only looked good when the terminal uses
a dark background. Now, it is possible to set the environment variable
"BACKGROUND", depending on its value an appropriate color is chosen.

In the future I want to add an automatic detection whether a dark or
light background is used by the terminal. That way I don't have to
specify the environment variable by hand.