emacs: introduce Org-roam
I heard a lot of good things about Org-roam. Today I finally tried it
and it made a good first impression on me.

While reading the entire documentation of Org-roam, I adjusted the
configuration based on the insights I gathered. For example the
`display-buffer-alist` configuration is heavily inspired by the
suggestions made in the documentation.

The user option `org-roam-directory` is not set here because it is
system-dependent. Because of this, it is also not possible to directly
enable `org-roam-db-autosync-mode` in this configuration as the Org-roam
directory has to be set before this command can be called.