emacs: customize user options using `setopt`
Since Emacs 29 it is recommended to use `setopt` instead of `setq` to
customize user options. This ensures that any associated `custom-set`
form of the user option is executed.

Since `setopt` is not as efficient as `setq`, I still use `setq` in the
function `mz/scroll-half-window-height`, as it is called very
frequently.