emacs: use `setq-local` to set buffer-local variables
In commit 3fe23c4 (emacs: customize user options using `setopt`,
2025-03-14), I used `setopt` in some hooks, not realizing that it
behaves like `setq-default` and modifies the default value instead of
the buffer-local value. [1] I noticed this when suddenly `fill-column`
was set to 72 globally, although this value should only be used in
`git-commit-mode`.

Since all variables in question are marked as buffer-local, using `setq`
would suffice to set their values. However, `setq-local` is used to
indicate that the changes are buffer-local, reducing potential confusion
in the future.

[1]: https://emacs.stackexchange.com/a/78424